+++ engine/src/main/java/net/jqwik/engine/execution/CheckedProperty.java
@@ -2,6 +2,7 @@ package net.jqwik.engine.execution;
 
 import java.util.*;
 import java.util.function.*;
+import java.util.logging.*;
 import java.util.stream.*;
 
 import net.jqwik.api.*;
@@ -13,6 +14,8 @@ import net.jqwik.engine.properties.*;
 import net.jqwik.engine.support.*;
 
 public class CheckedProperty {
+
+	private static final Logger LOG = Logger.getLogger(CheckedProperty.class.getName());
 
 	public final String propertyName;
 	public final TryLifecycleExecutor tryLifecycleExecutor;
@@ -27,14 +30,14 @@ public class CheckedProperty {
 	private Optional<ExhaustiveShrinkablesGenerator> optionalExhaustive;
 
 	public CheckedProperty(
-		String propertyName,
-		TryLifecycleExecutor tryLifecycleExecutor,
-		List<MethodParameter> propertyParameters,
-		ArbitraryResolver arbitraryResolver,
-		ResolveParameterHook resolveParameterHook,
-		PropertyLifecycleContext propertyLifecycleContext,
-		Optional<Iterable<? extends Tuple>> optionalData,
-		PropertyConfiguration configuration
+			String propertyName,
+			TryLifecycleExecutor tryLifecycleExecutor,
+			List<MethodParameter> propertyParameters,
+			ArbitraryResolver arbitraryResolver,
+			ResolveParameterHook resolveParameterHook,
+			PropertyLifecycleContext propertyLifecycleContext,
+			Optional<Iterable<? extends Tuple>> optionalData,
+			PropertyConfiguration configuration
 	) {
 		this.propertyName = propertyName;
 		this.tryLifecycleExecutor = tryLifecycleExecutor;
@@ -53,6 +56,7 @@ public class CheckedProperty {
 
 	public PropertyCheckResult check(Reporting[] reporting) {
 		PropertyConfiguration effectiveConfiguration = configurationWithEffectiveSeed();
+		maybeWarnOnMultipleTriesWithoutForallParameters(effectiveConfiguration);
 		try {
 			Reporter reporter = propertyLifecycleContext.reporter();
 			return createGenericProperty(effectiveConfiguration).check(reporter, reporting);
@@ -63,6 +67,19 @@ public class CheckedProperty {
 					configuration.getEdgeCasesMode(), 0, 0,
 					null, null, cannotFindArbitraryException
 			);
+		}
+	}
+
+	private void maybeWarnOnMultipleTriesWithoutForallParameters(PropertyConfiguration effectiveConfiguration) {
+		if (effectiveConfiguration.getTries() > 1 && forAllParameters.isEmpty()) {
+			String message = String.format(
+					"Running %s [%s] in container [%s] without any @ForAll parameters for %s tries.%n    Maybe you want to change it into an @Example?",
+					effectiveConfiguration.getStereotype(),
+					propertyLifecycleContext.extendedLabel(),
+					propertyLifecycleContext.containerClass().getName(),
+					effectiveConfiguration.getTries()
+			);
+			LOG.warning(message);
 		}
 	}
 
@@ -85,17 +102,17 @@ public class CheckedProperty {
 		} else if (configuration.getGenerationMode() == GenerationMode.EXHAUSTIVE) {
 			ensureValidExhaustiveMode();
 			configuration = configuration.withTries(
-				Math.toIntExact(getOptionalExhaustive().get().maxCount())
+					Math.toIntExact(getOptionalExhaustive().get().maxCount())
 			);
 		} else if (configuration.getGenerationMode() == GenerationMode.AUTO) {
 			configuration = chooseGenerationMode(configuration);
 		}
 		ForAllParametersGenerator shrinkablesGenerator = createShrinkablesGenerator(configuration);
 		ResolvingParametersGenerator parametersGenerator = new ResolvingParametersGenerator(
-			propertyParameters,
-			shrinkablesGenerator,
-			resolveParameterHook,
-			propertyLifecycleContext
+				propertyParameters,
+				shrinkablesGenerator,
+				resolveParameterHook,
+				propertyLifecycleContext
 		);
 		Supplier<TryLifecycleContext> tryLifecycleContextSupplier = () -> new DefaultTryLifecycleContext(propertyLifecycleContext);
 		return new GenericProperty(propertyName, configuration, parametersGenerator, tryLifecycleExecutor, tryLifecycleContextSupplier);
@@ -108,7 +125,7 @@ public class CheckedProperty {
 				return createSampleOnlyShrinkableGenerator(configuration);
 			} else if (configuration.getAfterFailureMode() == AfterFailureMode.SAMPLE_FIRST) {
 				return createSampleOnlyShrinkableGenerator(configuration)
-						   .andThen(() -> createDefaultShrinkablesGenerator(configuration));
+							   .andThen(() -> createDefaultShrinkablesGenerator(configuration));
 			}
 		}
 		return createDefaultShrinkablesGenerator(configuration);
@@ -163,7 +180,7 @@ public class CheckedProperty {
 		}
 		try {
 			ExhaustiveShrinkablesGenerator exhaustiveShrinkablesGenerator =
-				ExhaustiveShrinkablesGenerator.forParameters(forAllParameters, arbitraryResolver, maxNumberOfSamples);
+					ExhaustiveShrinkablesGenerator.forParameters(forAllParameters, arbitraryResolver, maxNumberOfSamples);
 			return Optional.of(exhaustiveShrinkablesGenerator);
 		} catch (TooManyFilterMissesException tmfme) {
 			throw tmfme;
@@ -182,11 +199,11 @@ public class CheckedProperty {
 	private ForAllParametersGenerator createRandomizedShrinkablesGenerator(PropertyConfiguration configuration) {
 		Random random = SourceOfRandomness.create(configuration.getSeed());
 		return RandomizedShrinkablesGenerator.forParameters(
-			forAllParameters,
-			arbitraryResolver,
-			random,
-			configuration.getTries(),
-			configuration.getEdgeCasesMode()
+				forAllParameters,
+				arbitraryResolver,
+				random,
+				configuration.getTries(),
+				configuration.getEdgeCasesMode()
 		);
 	}
 
@@ -198,7 +215,7 @@ public class CheckedProperty {
 		//noinspection OptionalAssignedToNull
 		if (optionalExhaustive == null) {
 			long maxNumberOfSamples = configuration.getGenerationMode() == GenerationMode.EXHAUSTIVE
-										  ? ExhaustiveGenerator.MAXIMUM_SAMPLES_TO_GENERATE : configuration.getTries();
+											  ? ExhaustiveGenerator.MAXIMUM_SAMPLES_TO_GENERATE : configuration.getTries();
 			optionalExhaustive = createOptionalExhaustiveShrinkablesGenerator(maxNumberOfSamples);
 		}
 		return optionalExhaustive;