+++ engine/src/test/java/examples/packageWithErrors/ContainerWithOOME.java
@@ -5,7 +5,7 @@ import net.jqwik.api.*;
 public class ContainerWithOOME {
 
 	@Property(tries = 10)
-	void throwOutOfMemoryError() {
+	void throwOutOfMemoryError(@ForAll int i) {
 		throw new OutOfMemoryError();
 	}
 }