+++ engine/src/test/java/examples/packageWithSeveralContainers/PropertyTests.java
@@ -1,7 +1,9 @@
 package examples.packageWithSeveralContainers;
 
 import net.jqwik.api.*;
+import net.jqwik.testing.*;
 
+@SuppressLogging
 public class PropertyTests {
 
 	@Property