+++ engine/src/test/java/net/jqwik/engine/execution/lifecycle/AroundTryHookTests.java
@@ -9,6 +9,7 @@ import net.jqwik.testing.*;
 
 import static org.assertj.core.api.Assertions.*;
 
+@SuppressLogging
 class AroundTryHookTests {
 
 	static int count1 = 0;