+++ engine/src/test/java/net/jqwik/engine/properties/ResolvingParametersInTryTests.java
@@ -10,6 +10,7 @@ import net.jqwik.testing.*;
 
 import static org.assertj.core.api.Assertions.*;
 
+@SuppressLogging
 class ResolvingParametersInTryTests {
 
 	@Property(tries = 10)